switched back upcomming epoch parameter type from uint32 to uint256: it dit not save up any memory space, it just added additional problems and confusion.

+++ contracts/KeyGenHistory.sol
@@ -93,7 +93,7 @@ contract KeyGenHistory is UpgradeabilityAdmin, IKeyGenHistory {
         }
     }
 
-    function writePart(uint32 _upcommingEpoch, bytes memory _part)
+    function writePart(uint256 _upcommingEpoch, bytes memory _part)
     public
     onlyUpcommingEpoch(_upcommingEpoch) {
         // It can only be called by a new validator which is elected but not yet finalized...
